requirements.txt
quixstreams
requests
pandas
redis
python-dotenv
pytrends
scikit-learn
streamlit